Under construction page added
Added an under construction page for windows with a width under 1254px. This is where some buttons and elements break up.

The Instagram SVG is now as big as the others. Fixed this by removing the image within the circle and adding a new gradient in Illustrator.
The smooth linking from a project page to the projects section on the index page is not working anymore, probably because of the JS.